server/requirements.txt
# FastAPI server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Core ML dependencies
transformers>=4.53.0
torch>=2.1.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
safetensors>=0.3.0

# Multimodal support
Pillow>=9.0.0
opencv-python>=4.8.0
librosa>=0.10.0
moviepy>=1.0.3

# Additional utilities
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=1.0.0
aiofiles>=23.0.0

# OpenAI compatibility and LangChain integration
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Development and monitoring
httpx>=0.25.0
psutil>=5.9.0

# Additional dependencies
sse-starlette>=1.6.0
aiohttp>=3.8.0 